Map & seq encoding by mime type.
Allow custom encoders to be specified in a thread-safe, per-mime-type
manner. Encoders are kept in a map keyed by mime type and wrapped in
a dynamic var. When `install-{map,seq}-representation-dispatcher` is
caled, it adds a method to liberator's representation multimethod for
the given mime type that looks in the dynamic var to try to find an
encoder for the mime type, and falls back to the provided default if
one could not be found.

There are also functions that, given a mime type and an encoder,
generate middleware that sets that encoder to be used for that mime type
for the handler that the middleware wraps.